==> 5_sentence_transformer.ipynb <==
# REF: https://huggingface.co/blog/getting-started-with-embeddings
# REF: https://www.sbert.net/

import pandas as pd

# load grouped data
src_train = 'data/training_data_grouped.csv'
training_data_raw = pd.read_csv(src_train)

src_val = 'data/validation_data_grouped.csv'
validation_data_raw = pd.read_csv(src_val)

src_test = 'data/test_data_grouped.csv'
test_data_raw = pd.read_csv(src_test)

src_liar = 'data/liar_dataset/test_features.csv'
liar_data_raw = pd.read_csv(src_liar)

# get texts
type = 'content_clean'
txt_train = training_data_raw[type]
txt_val = validation_data_raw[type]
txt_test = test_data_raw[type]
txt_liar = liar_data_raw[type]

from sentence_transformers import SentenceTransformer

# load transformer model
# see models: https://www.sbert.net/docs/pretrained_models.html#model-overview
model = SentenceTransformer("all-MiniLM-L12-v2")

# get encoded output from transformer
# TIME: about 22min on M1 MacBook Pro
train_encodings = model.encode(txt_train)
val_encodings = model.encode(txt_val)
test_encodings = model.encode(txt_test)
liar_encodings = model.encode(txt_liar)

# write to csv file
dst = 'data/training_data_embeddings.csv'
embeddings = pd.DataFrame(train_encodings)
embeddings.to_csv(dst, index=False)

dst = 'data/validation_data_embeddings.csv'
embeddings = pd.DataFrame(val_encodings)
embeddings.to_csv(dst, index=False)

dst = 'data/test_data_embeddings.csv'
embeddings = pd.DataFrame(test_encodings)
embeddings.to_csv(dst, index=False)

dst = 'data/liar_dataset/test_embeddings.csv'
embeddings = pd.DataFrame(liar_encodings)
embeddings.to_csv(dst, index=False)
